[chore] Session에 저장할 SessionEmployeeDTO 객체 추가 및 메일 기능 추가

+++ src/main/java/kr/co/groupworks/dto/cis/employee/EmployeeDTO.java
@@ -12,7 +12,7 @@ import lombok.experimental.Accessors;
 @NoArgsConstructor
 public class EmployeeDTO {
     //  사원번호 pk
-    private int employeeId;
+    private Long employeeId;
     //  사원명
     private String employeeName;
     //  직급 fk
